+++ api/tests/acceptance/application/assessment-controller-get-nonadaptive_test.js
@@ -60,7 +60,7 @@ describe('Acceptance | API | Assessments GET (non adaptive)', function () {
 
   describe('(non-adaptive) GET /api/assessments/:assessment_id/next', function () {
 
-    let inserted_assessment_id = null;
+    let insertedAssessmentId = null;
 
     const inserted_assessment = {
       userName: 'John Doe',
@@ -71,7 +71,7 @@ describe('Acceptance | API | Assessments GET (non adaptive)', function () {
     beforeEach(function (done) {
       knex('assessments').delete().then(() => {
         knex('assessments').insert([inserted_assessment]).then((rows) => {
-          inserted_assessment_id = rows[0];
+          insertedAssessmentId = rows[0];
           done();
         });
       });
@@ -82,7 +82,7 @@ describe('Acceptance | API | Assessments GET (non adaptive)', function () {
     });
 
     it('should return 200 HTTP status code', function (done) {
-      const options = { method: 'GET', url: '/api/assessments/' + inserted_assessment_id + '/next' };
+      const options = { method: 'GET', url: '/api/assessments/' + insertedAssessmentId + '/next' };
       server.inject(options, (response) => {
         expect(response.statusCode).to.equal(200);
         done();
@@ -90,7 +90,7 @@ describe('Acceptance | API | Assessments GET (non adaptive)', function () {
     });
 
     it('should return application/json', function (done) {
-      const options = { method: 'GET', url: '/api/assessments/' + inserted_assessment_id + '/next' };
+      const options = { method: 'GET', url: '/api/assessments/' + insertedAssessmentId + '/next' };
       server.inject(options, (response) => {
         const contentType = response.headers['content-type'];
         expect(contentType).to.contain('application/json');
@@ -99,7 +99,7 @@ describe('Acceptance | API | Assessments GET (non adaptive)', function () {
     });
 
     it('should return the first challenge if no challenge specified', function (done) {
-      const options = { method: 'GET', url: '/api/assessments/' + inserted_assessment_id + '/next' };
+      const options = { method: 'GET', url: '/api/assessments/' + insertedAssessmentId + '/next' };
       server.inject(options, (response) => {
         expect(response.result.data.id).to.equal('first_challenge');
         done();
@@ -107,7 +107,7 @@ describe('Acceptance | API | Assessments GET (non adaptive)', function () {
     });
 
     it('should return the next challenge otherwise', function (done) {
-      const options = { method: 'GET', url: '/api/assessments/' + inserted_assessment_id + '/next/first_challenge' };
+      const options = { method: 'GET', url: '/api/assessments/' + insertedAssessmentId + '/next/first_challenge' };
       server.inject(options, (response) => {
         expect(response.result.data.id).to.equal('second_challenge');
         done();
@@ -115,7 +115,15 @@ describe('Acceptance | API | Assessments GET (non adaptive)', function () {
     });
 
     it('should return null if reached the last challenge of the course', function (done) {
-      const options = { method: 'GET', url: '/api/assessments/' + inserted_assessment_id + '/next/second_challenge' };
+      const options = { method: 'GET', url: '/api/assessments/' + insertedAssessmentId + '/next/second_challenge' };
+      server.inject(options, (response) => {
+        expect(response.result).to.equal('null');
+        done();
+      });
+    });
+
+    it('should return null if reached the last challenge of the course', function (done) {
+      const options = { method: 'GET', url: '/api/assessments/' + insertedAssessmentId + '/next/second_challenge' };
       server.inject(options, (response) => {
         expect(response.result).to.equal('null');
         done();
