+++ api/tests/acceptance/application/assessment-controller-get-solutions_test.js
@@ -8,6 +8,7 @@ describe('Acceptance | API | Solutions of Assessments', function () {
       knex.seed.run().then(() => {
         nock('https://api.airtable.com')
           .get('/v0/test-base/Tests/non_adaptive_course_id')  // XXX cf. issue #204, there may be a conflict with course-controller_test
+          .query(true)
           .times(4)
           .reply(200, {
             'id': 'non_adaptive_course_id',
@@ -23,6 +24,7 @@ describe('Acceptance | API | Solutions of Assessments', function () {
           );
         nock('https://api.airtable.com')
           .get('/v0/test-base/Tests/adaptive_course_id')
+          .query(true)
           .times(4)
           .reply(200, {
             'id': 'adaptive_course_id',
@@ -39,6 +41,7 @@ describe('Acceptance | API | Solutions of Assessments', function () {
           );
         nock('https://api.airtable.com')
           .get('/v0/test-base/Epreuves/first_challenge')
+          .query(true)
           .times(3)
           .reply(200, {
             'id': 'first_challenge',
@@ -49,6 +52,7 @@ describe('Acceptance | API | Solutions of Assessments', function () {
           );
         nock('https://api.airtable.com')
           .get('/v0/test-base/Epreuves/second_challenge')
+          .query(true)
           .reply(200, {
             'id': 'second_challenge',
             'fields': {
@@ -58,6 +62,7 @@ describe('Acceptance | API | Solutions of Assessments', function () {
           );
         nock('https://api.airtable.com')
           .get('/v0/test-base/Epreuves/third_challenge')
+          .query(true)
           .reply(200, {
             'id': 'third_challenge',
             'fields': {
